Note Simple Registry, Thankful, and article on registries.
The takeaway from the article is that none of the Universal registry
solutions are a magic bullet. Simple Registry might be a good one to use
just for things that we are cool with collecting cash on.

Thankful might be for those things that we want people to actually
complete the purchase process for.